+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.9.0 - 2026-01-12
+
+- Added **Topic Test** feature, will add documentation for Supabase in next minor version.
+- Added animated API support.
+- Removed ghost API calls.
+- Fixed bug which prevented changing name.
+- Added [Dexie](https://dexie.org/) for indexedDB and for future local-first solution.
+
 # v0.8.0 – 2025-12-16
 
 - Added **Smart Revision** feature. [See documentation](./SUPABASE_DOCUMENTATION.md)